# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Locadora de Carros",
        "description": "Documentação da API da Locadora de Carros com autenticação JWT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/v1/locacao": {
            "get": {
                "tags": [
                    "Locacao"
                ],
                "summary": "Locacao cadastradas",
                "operationId": "92b8080c5793f4f3f6e7c6bbee29bc4b",
                "parameters": [
                    {
                        "name": "atributos_clientes",
                        "in": "query",
                        "description": "Quais colunas do relacionamento 'cliente' retornar. Use vírgula para separar.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "nome"
                    },
                    {
                        "name": "atributos_carros",
                        "in": "query",
                        "description": "Quais colunas do relacionamento 'carro' retornar. Use vírgula para separar.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "placa,disponivel,km"
                    },
                    {
                        "name": "filtro",
                        "in": "query",
                        "description": "Filtros dinâmicos. Exemplos de operadores comuns: =, >, >=, <, <=, like, in, between. ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "km_final:>:1000"
                    },
                    {
                        "name": "atributos",
                        "in": "query",
                        "description": "Quais colunas do recurso 'locacao' retornar. Use vírgula para separar.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "cliente_id,carro_id,data_inicio_periodo,data_final_previsto_periodo,km_final"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os dados da locação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Locacao"
                ],
                "summary": "Cadastro de locações",
                "operationId": "2e36625b60e2ac69bb6653ca6c137e12",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "carro_id",
                                    "data_inicio_periodo",
                                    "data_final_previsto_periodo",
                                    "data_final_realizado_periodo",
                                    "valor_diaria"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "carro_id": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "data_inicio_periodo": {
                                        "type": "date_format:d-m-y",
                                        "example": "18-08-25"
                                    },
                                    "data_final_previsto_periodo": {
                                        "type": "date_format:d-m-y",
                                        "example": "20-08-25"
                                    },
                                    "data_final_realizado_periodo": {
                                        "type": "date_format:d-m-y",
                                        "example": "20-08-25"
                                    },
                                    "valor_diaria": {
                                        "type": "integer",
                                        "example": "150"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Locação criada com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/locacao/{id}": {
            "get": {
                "tags": [
                    "Locacao"
                ],
                "summary": "Detalhes de uma locacao",
                "operationId": "b604e242a8c756b6200372eb0ff360f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da locacao a ser pesquisado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Locacao não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Locacao"
                ],
                "summary": "Atualização completa de uma locacao",
                "description": "Substitui todos os campos do recurso locacao (estilo PUT).",
                "operationId": "08213c4c51f3867e134052b8759948b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da locacao a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Payload completo para atualização (PUT requer todos os campos obrigatórios do model).",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "carro_id",
                                    "data_inicio_periodo",
                                    "data_final_previsto_periodo",
                                    "valor_diaria"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "carro_id": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "data_inicio_periodo": {
                                        "type": "date_format:d-m-y",
                                        "example": "18-08-25"
                                    },
                                    "data_final_previsto_periodo": {
                                        "type": "date_format:d-m-y",
                                        "example": "20-08-25"
                                    },
                                    "valor_diaria": {
                                        "type": "integer",
                                        "example": "150"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carro atualizado com sucesso"
                    },
                    "404": {
                        "description": "Recurso não existe"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Locacao"
                ],
                "summary": "Remover uma locacao",
                "operationId": "c322668d896d6c467943b879e3963bd4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da locacao a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Locacao não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Locacao"
                ],
                "summary": "Finaliza uma locacao",
                "description": "Atualiza apenas os campos enviados (estilo PATCH).",
                "operationId": "422ae94da5bfa9c1fc2a97fc0d4440bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da locacao a ser atualizado parcialmente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Qualquer subconjunto de campos de locacao.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "km_final": {
                                        "type": "integer",
                                        "example": 1798
                                    },
                                    "data_final_realizado_periodo": {
                                        "type": "date_format:d-m-y",
                                        "example": "20-08-25"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Locacao atualizado parcialmente com sucesso"
                    },
                    "404": {
                        "description": "Recurso não existe"
                    },
                    "422": {
                        "description": "Erro de validação (apenas para campos enviados)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT no formato: Bearer {seu_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}